+++ src/main/java/com/winner/trelloimplementation/user/controller/HomeController.java
@@ -1,0 +1,39 @@
+package com.winner.trelloimplementation.user.controller;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.winner.trelloimplementation.common.jwt.JwtUtil;
+import com.winner.trelloimplementation.user.service.KakaoService;
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+@Controller
+@RequiredArgsConstructor
+@RequestMapping("/api")
+public class HomeController {
+
+    private final KakaoService kakaoService;
+
+    @GetMapping("/sign")
+    public String signPage() { return "sign"; }
+
+    @GetMapping("/my-page")
+    public String myPage() { return "my-page"; }
+
+    @GetMapping("/user/kakao/callback")
+    public String kakaoLogin(@RequestParam String code, HttpServletResponse response) throws JsonProcessingException {
+        // code: 카카오 서버로부터 받은 인가 코드 Service 전달 후 인증 처리 및 JWT 반환
+        String token = kakaoService.kakaoLogin(code);
+
+        // Cookie 생성 및 직접 브라우저에 Set
+        Cookie cookie = new Cookie(JwtUtil.AUTHORIZATION_HEADER, token.substring(7));
+        cookie.setPath("/");
+        response.addCookie(cookie);
+
+        return "redirect:/board";
+    }
+}
