+++ src/main/java/com/winner/trelloimplementation/user/controller/UserController.java
@@ -1,11 +1,8 @@
 package com.winner.trelloimplementation.user.controller;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.winner.trelloimplementation.common.dto.ApiResponseDto;
-import com.winner.trelloimplementation.common.jwt.JwtUtil;
 import com.winner.trelloimplementation.common.security.UserDetailsImpl;
 import com.winner.trelloimplementation.user.dto.*;
-import com.winner.trelloimplementation.user.service.KakaoService;
 import com.winner.trelloimplementation.user.service.UserServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -14,11 +11,9 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.annotation.Nullable;
 import jakarta.servlet.ServletException;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -32,8 +27,6 @@ import java.io.IOException;
 @RequestMapping("/api/user")
 public class UserController {
     private final UserServiceImpl userServiceImpl;
-    private final JwtUtil jwtUtil;
-    private final KakaoService kakaoService;
 
     @Operation(summary = "로그인 메서드", description = "로그인 메서드입니다.")
     @ApiResponses(value = {
@@ -108,18 +101,5 @@ public class UserController {
     public ResponseEntity<ApiResponseDto> signout(@AuthenticationPrincipal UserDetailsImpl userDetails, @RequestBody SignoutRequestDto signoutRequestDto,
                                                   HttpServletResponse response, Authentication authResult) throws ServletException, IOException {
         return userServiceImpl.signout(userDetails.getUser(), signoutRequestDto, response, authResult);
-    }
-
-    @GetMapping("/user/kakao/callback")
-    public String kakaoLogin(@RequestParam String code, HttpServletResponse response) throws JsonProcessingException {
-        // code: 카카오 서버로부터 받은 인가 코드 Service 전달 후 인증 처리 및 JWT 반환
-        String token = kakaoService.kakaoLogin(code);
-
-        // Cookie 생성 및 직접 브라우저에 Set
-        Cookie cookie = new Cookie(JwtUtil.AUTHORIZATION_HEADER, token.substring(7));
-        cookie.setPath("/");
-        response.addCookie(cookie);
-
-        return "redirect:/";
     }
 }
