+++ src/main/resources/static/js/sign.js
@@ -1,0 +1,5 @@
+document.querySelector('.img-btn').addEventListener('click', function()
+    {
+        document.querySelector('.cont').classList.toggle('s-signup')
+    }
+);
