--- src/main/resources/templates/login.html
@@ -1,79 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport"
-        content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <link rel="preconnect" href="https://fonts.gstatic.com">
-  <link rel="stylesheet" type="text/css" href="/css/style.css">
-  <script src="https://code.jquery.com/jquery-3.7.0.min.js"
-          integrity="sha256-2Pmvv0kuTBOenSvLm6bvfBSSHrUJ+3A7x6P5Ebd07/g=" crossorigin="anonymous"></script>
-  <script src="https://cdn.jsdelivr.net/npm/js-cookie@3.0.5/dist/js.cookie.min.js"></script>
-  <meta charset="UTF-8">
-  <title>로그인 페이지</title>
-</head>
-<body>
-<div id="login-form">
-  <div id="login-title">Log into Select Shop</div>
-  <button id="login-kakao-btn" onclick="location.href='https://kauth.kakao.com/oauth/authorize?client_id=3a94933819924124b9c1cd5213aa1759&redirect_uri=http://localhost:8080/api/user/kakao/callback&response_type=code'">
-    카카오로 로그인하기
-  </button>
-  <button id="login-id-btn" onclick="location.href='/api/user/signup'">
-    회원 가입하기
-  </button>
-  <div>
-    <div class="login-id-label">아이디</div>
-    <input type="text" name="username" id="username" class="login-input-box">
-
-    <div class="login-id-label">비밀번호</div>
-    <input type="password" name="password" id="password" class="login-input-box">
-
-    <button id="login-id-submit" onclick="onLogin()">로그인</button>
-  </div>
-  <div id="login-failed" style="display:none" class="alert alert-danger" role="alert">로그인에 실패하였습니다.</div>
-</div>
-</body>
-<script>
-  $(document).ready(function () {
-    // 토큰 삭제
-    Cookies.remove('Authorization', {path: '/'});
-  });
-
-  const href = location.href;
-  const queryString = href.substring(href.indexOf("?") + 1)
-  if (queryString === 'error') {
-    const errorDiv = document.getElementById('login-failed');
-    errorDiv.style.display = 'block';
-  }
-
-  const host = 'http://' + window.location.host;
-
-  function onLogin() {
-    let username = $('#username').val();
-    let password = $('#password').val();
-
-    $.ajax({
-      type: "POST",
-      url: `/api/user/login`,
-      contentType: "application/json",
-      data: JSON.stringify({username: username, password: password}),
-    })
-            .done(function (res, status, xhr) {
-              const token = xhr.getResponseHeader('Authorization');
-
-              Cookies.set('Authorization', token, {path: '/'})
-
-              $.ajaxPrefilter(function (options, originalOptions, jqXHR) {
-                jqXHR.setRequestHeader('Authorization', token);
-              });
-
-              window.location.href = host;
-            })
-            .fail(function (jqXHR, textStatus) {
-              alert("Login Fail");
-              window.location.href = host + '/api/user/login-page?error'
-            });
-  }
-</script>
-</html>
